Удалил ненужные методы, подправил баг с некопированной картинкой при расчете подкартинки с использованием opencv, написал подробные комментарии для алгоритмов.

+++ Correction/Correction/NodesSet.cpp
@@ -126,7 +126,7 @@ bool NodesSet::empty()
 	return nodes_.empty();
 }
 
-float NodesSet::getAngle(int row, int col, int dir) 
+double NodesSet::getAngle(int row, int col, int dir) 
 {
 	assert(!empty());
 	assert(row >= 0 && row < rows() && col >= 0 && col < cols());
